Finished Membership

+++ database/migrations/2025_09_22_132933_create_memberships_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('membership', function (Blueprint $table) {
             $table->id();
             $table->float('harga');
-            $table->bigInteger('durasi');
+            $table->integer('durasi');
             $table->string('nama_plan');
             $table->string('deskripsi');
             $table->timestamps();
